Tests/LibWeb/Text/expected/selection-drag-above-viewport-with-same-row-sidebar.txt

selected main title: true
selected forward text: false
selected sidebar text: false
